requirements.txt
asgiref==3.8.1
click==8.1.7
cmake==3.31.0.1
colorama==0.4.6
Django==5.1.3
dlib==19.24.1
dnspython==2.7.0
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.16.1
idna==3.10
keyboard==0.13.5
numpy @ file:///D:/bld/numpy_1730588036902/work/dist/numpy-2.1.3-cp311-cp311-win_amd64.whl#sha256=102c9d16b55bb9d4d150a9fc7d35c3c22b00774791127daaf0d375a7ca3464be
opencv-python==4.10.0.84
pillow==11.0.0
py3-validate-email==1.0.5.post2
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
validate-email-address==1
